+++ CHANGES.md
@@ -2,8 +2,8 @@
 
 ## Unreleased - 2026-08-04
 
-Changes since v0.6.4.
+Changes since v0.6.5.
 
 ### 🐛 Bug Fixes
 
-- Password reset and logout for email/password accounts (1cbe35f).
+- User preferences (cebd9f0).